pyproject.toml
[project]
name = "blogify-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.3",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "youtube-transcript-api>=1.2.3",
]
